Eric Sunshine made the following changes:
-*- Eric Sunshine updated out-of-date mk/autoconf/*.m4 files, including
    the new crystal.m4 which now published CRYSTAL.INCLUDE_DIR.

-*- Eric Sunshine fixed behaviorlayer/python/Jamfile so that it passes
    only -I$(CRYSTAL.INCLUDE_DIR) to Swig, rather than all of
    $(CRYSTAL.CFLAGS), which was clearly incorrect, and which caused Swig
    to complain in some cases.

-*- Eric Sunshine eliminated several compilation warnings from the
    project.

+++ plugins/tools/billboard/billboard.cpp
@@ -94,8 +94,8 @@ void celBillboard::TranslateScreenToTexture (int sx, int sy, int& tx, int& ty)
     if (ty >= image_h) ty = image_h-1;
   }
 
-  tx = tx * (uv_botright.x - uv_topleft.x) + uv_topleft.x * image_w;
-  ty = ty * (uv_botright.y - uv_topleft.y) + uv_topleft.y * image_h;
+  tx = (int)(tx * (uv_botright.x - uv_topleft.x) + uv_topleft.x * image_w);
+  ty = (int)(ty * (uv_botright.y - uv_topleft.y) + uv_topleft.y * image_h);
 }
 
 bool celBillboard::GetFromClickMap (int tx, int ty)
